+++ init.lua
@@ -143,6 +143,8 @@ vim.lsp.config("roslyn", {
 vim.lsp.enable('pylsp')
 vim.lsp.enable('rust_analyzer')
 vim.lsp.enable('lua_ls')
+vim.lsp.enable('ts_ls')
+vim.lsp.enable('astro')
 vim.lsp.enable('asm_lsp')
 vim.lsp.enable('bashls')
 vim.lsp.enable('roslyn')